Change options argument passed over to simple-git push to an object
This method has a very confusing signature. The comment in the source-code suggests you can pass all arguments as an options array as part of the 1st argument (remote) and that's what I was doing. However, it's not compatible with the type definitions. You can't pass an array of options as the last arguments as it will just swallow them silently. For the last argument to be recognized as git options, you should pass it as a JS object. If the options does not accept arguments, its value should be null.

Changing it to an object fixed the type errors too.